django-post/requirements.txt
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
Django==2.1.5
django-extensions==2.1.4
djangorestframework==3.9.1
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.2
Jinja2==2.10
MarkupSafe==1.1.0
parso==0.3.1
pickleshare==0.7.5
prompt-toolkit==2.0.7
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
